✨ [Feat] TodoModel 타입 변환을 위한 확장함수 추가
